requirements.txt
fastapi>=0.100.0
pymongo==4.5.0
pydantic==1.10.2
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
uvicorn==0.23.2